class, function, object
Class is a conatianer which contains :-
1) data members
2)member functions,
3)constructor,
4)static blocks,
5)static data memebers,
6)static functions,
7)nested or inner class.

1 and 2 are called with object
3 is called when object is created
4 and 5  are called with classname
7  are called with operator

Imp. Points:-
1) We can declare variable in class but can't initialize them
2) Not even print statements can execute in it.
3) We can initialize variables and use print statements in a block,
function, constructor
4) Each class has a default constructor

Functions are name block which are used to define the functionality for
data memebers. Functions are used to perform the specific task. There
are pre-defined and user defined functions. Pre-defined functions are
those which are already defined in java library. User defined functions
are those functions which are used to perform specific task by the user.
There are four type of functions:-
1) Function with no argument and no return value.
2) Function with argument and no return value.
3) Function with argument and return value.
4) Function with no argument and return value.

Rules of function :-
1) always start with small letter
2) function should have name and return type
3) For return type we can use premitive data type, void, classes, array,
premitive class object.
4) Access specifier are optional but we can use only one at a time.
5) We can use multiple modifiers with function.
6) We should have specify the parameters according to the return type of
the function.
7) Function can return only one value.

Format:-
accessSpecifier modifier returnType functionName(parameters){}

Object is an instance of a class. When we create a object, it create
copy of data member separately. Objects are stored in heap.

Format:-
className objectName = new className();

Note:- Variables are stored in stack
Note:- new operator is used to allocated memory in the heap.

Function with no argument and no return type is used to display the
message.

Format:-
void functionName(){}

Function with argument and no return type is used to initialize the data
members of a class.

Format:-
void functionName(parameters){}

There are two type of arguments. They are Actual and Formal arguments.
Actual arguments are those which are used during calling a function.
Formal arguments are those which catches the argument in a function.

Function with argument and return value is used for input and output
purpose. In the function body we have to specify the return statement.

Format:-
public returnType functionName(parameters){}