this keyword, classes
using this keyword to pass arguments within the same class and to
different class.
Note:- to use the this keyword, we have to use it within the class
because this reprents the object of that class
for passing argument to different class we need create object of that
class in the class consisting of the actual variables.

Types of classes :-
1) Simple class
2) public class
3) final class
4) inner class
5) nested class
6) static class
7) abstract class
8) anonymus class
9)concrete class

Simple class is that we normally create

Public class can be only one and file name should be same as of class.
Public keyword is used with class name.

Final Class is used to restrict the inherritance but we can create
object of it. Final keyword is used with class name.

Inner class are those which are declared inside the class. We can
specify inner class as static or non static.

Static class created inside the class is known as static nested class.
It can't access non static data member or method. It can be accessed by
outer class name. It can access static member of outer class including
private.

Note:- We can't create any simple class as static
Points for inner class
1) The inner class can access the variable of outer class.
2) We have to create object of outer class before accessing the member
of inner class
Format :-
outerClass.InnerClass obj_name = objOfOuterClass.new Innerclass();

Points of Static nested class:-
1) The static nested class will access only static data members of outer
class
2) We don't need to declare the object of outer class to create the
object of static nested class
Format:-
outerClass.InnerClass obj_name = new outerClass.InnerClass();