Polymorphism
Polymorphism is a process in which we specify function name same but the
parameters are different. It provide feature of reducing name of the
function. we can use it within the same class. It improves he
performance of compiler. There are two type of Polymorphism.
1) Compile-time
2) Run-time

Compile-time is that in which we have function name same but the
parameters are different. Function overloading and constructor
overloading are the examples.

Points to remeber:-
1) Return Type doesn't matter
2) Function name always will be same
3) Data type can be same but the number of arguments must be different
4) Number of arguments are same but data type must be different.
5) Data type is same, number of argument is same but the order is
reversed.