inherritance
Inherritance is a process in which we can inherrit the property and
method of parent in the child class. It provide the facility of code
reusability.

Keypoints:-
1) Inherritance use extends keyword.
2) Constructors Can't be inherrited.
3) Private members can't inherrited.
4) Final class con't inherrited
5) Final method can be inherrited but can't be overrided.
6) In inherritance we have to create object of child class.
7) Default constructor of base in Inherritace is called by the child
class.
8) We need to define explicitly default constructor if we are defining
parameterized constructor.
9) If we are using super keyword in child class to called forcebly the
parameterized constructor of base class then we don't need to define
default constructor explicitly.

Super is a keyword which is used call immideate parent class
constructor. We can use super with variable and method.

Overriding :- when we have method name and parameters same that is
considered as overriding.