Applets
Applets are small program which is used for graphical programming. There
are two type of applet, we can use in Java
1) Local Applet
2) Remote Applet

Local applet are those applet which are created in local machine and
executed on same machine.

Remote applets are those applets which are stored on the remote machine.
We can download them and execute on the local machine. To execute the
applet, we need java enabled browser.

LifeCycle of Applet:-
Normal applet contain lifecycle. During it's lifecycle it executes 5
methods. Those methods are:-
1) init()
2) start()
3) stop()
4) destroy()
5) paint(Grapiphics g)

init() executes once during lifecycle of a applet. It is used to
initialize the objects and variables which are used in program.

start() is used to start the applet. This method called automatically
when applet is executed.

stop() is used to stop the applet. This methos is also called
automatically. When the stop the applet but don't remove from memory.

destroy() is used to destroy the applet. This method is also called
automatically. This method removes the applet from the memory.

paint(Graphics g) is used to display the contents on the applet. To
display the content we use Graphics class method.